refactor: Modularized the API's source code
