+++ api/src/config/constants.ts
@@ -1,0 +1,3 @@
+export const CACHE_TTL = 3600000;
+export const SCHEMA_VERSION = 'v2';
+export const DEFAULT_CDN_URL = `https://github.com/eliasbenb/PlexAniBridge-Mappings/raw/refs/heads/${SCHEMA_VERSION}/mappings.json`;
