+++ api/src/routes/index.ts
@@ -1,0 +1,2 @@
+export { GetAllMappings } from './getAllMappings';
+export { SearchMappings } from './searchMappings';
