+++ api/src/types/index.ts
@@ -1,0 +1,219 @@
+import { z } from 'zod';
+
+import { Context } from 'hono';
+
+export type Env = {
+    CDN_URL: string;
+}
+
+export type AppContext = Context<{ Bindings: Env }>;
+
+export interface AnimeMappings {
+    [id: string]: AniMap;
+}
+
+export interface AniMap {
+    anidb_id: number | null;
+    anilist_id: number | null;
+    imdb_id: string | string[] | null;
+    mal_id: number | number[] | null;
+    tmdb_movie_id: number | number[] | null;
+    tmdb_show_id: number | number[] | null;
+    tvdb_id: number | null;
+    tvdb_mappings: Record<string, string> | null;
+}
+
+export interface Indexes {
+    anidb_id: Record<string, string[]>;
+    anilist_id: Record<string, string[]>;
+    imdb_id: Record<string, string[]>;
+    mal_id: Record<string, string[]>;
+    tmdb_movie_id: Record<string, string[]>;
+    tmdb_show_id: Record<string, string[]>;
+    tvdb_id: Record<string, string[]>;
+}
+
+export interface CacheEntry {
+    data: AnimeMappings;
+    indexes: Indexes;
+    timestamp: number;
+}
+
+export type Condition = {
+    field: "anidb_id" | "anilist_id" | "imdb_id" | "mal_id" | "tmdb_movie_id" | "tmdb_show_id" | "tvdb_id";
+    op: "eq" | "neq" | "in" | "nin" | "gt" | "gte" | "lt" | "lte";
+    value: string | number | (string | number)[];
+};
+
+export type LogicalFilter = { and: Filter[] } | { or: Filter[] };
+export type Filter = Condition | LogicalFilter;
+
+export const AniMapSchema = z.object({
+    anidb_id: z.number().nullable().openapi({
+        type: 'integer',
+        description: 'The matching AniDB ID for the entry',
+        example: 12820
+    }),
+    anilist_id: z.number().nullable().openapi({
+        type: 'integer',
+        description: 'The AniList ID of the anime',
+        example: 98310
+    }),
+    imdb_id: z.union([z.string(), z.array(z.string()), z.null()]).openapi({
+        oneOf: [
+            { type: 'string' },
+            { type: 'array', items: { type: 'string' } },
+            { type: 'null' }
+        ],
+        description: 'The matching IMDb ID(s) for the entry. Can be a single ID or an array of IDs',
+        example: 'tt9288776'
+    }),
+    mal_id: z.union([z.number(), z.array(z.number()), z.null()]).openapi({
+        oneOf: [
+            { type: 'integer' },
+            { type: 'array', items: { type: 'integer' } },
+            { type: 'null' }
+        ],
+        description: 'The matching MyAnimeList ID(s) for the entry. Can be a single ID or an array of IDs',
+        example: [34915, 36186, 36529]
+    }),
+    tmdb_movie_id: z.union([z.number(), z.array(z.number()), z.null()]).openapi({
+        oneOf: [
+            { type: 'integer' },
+            { type: 'array', items: { type: 'integer' } },
+            { type: 'null' }
+        ],
+        description: 'The matching TMDB ID(s) for the movie entry. Can be a single ID or an array of IDs',
+        example: 532067
+    }),
+    tmdb_show_id: z.union([z.number(), z.array(z.number()), z.null()]).openapi({
+        oneOf: [
+            { type: 'integer' },
+            { type: 'array', items: { type: 'integer' } },
+            { type: 'null' }
+        ],
+        description: 'The matching TMDB ID(s) for the show entry. Can be a single ID or an array of IDs',
+        example: 83095
+    }),
+    tvdb_id: z.number().nullable().openapi({
+        type: 'integer',
+        description: 'The matching TVDB ID for the show entry',
+        example: 333234
+    }),
+    tvdb_mappings: z.record(
+        z.string(),
+        z.string()
+    ).nullable().openapi({
+        type: 'object',
+        description: 'A dictionary mapping TVDB seasons to episode patterns. Keys are season identifiers in format "s0", "s1", etc. Values are episode patterns.',
+        example: {
+            's0': 'e1',
+            's1': 'e1-e13'
+        }
+    })
+
+}).openapi({
+    title: 'Anime Mapping',
+    description: 'Anime mapping data that connects IDs across different anime databases',
+    properties: {
+        anidb_id: {
+            type: 'integer',
+            nullable: true,
+            description: 'The matching AniDB ID for the entry',
+            example: 12820
+        },
+        anilist_id: {
+            type: 'integer',
+            nullable: true,
+            description: 'The AniList ID of the anime',
+            example: 98310
+        },
+        imdb_id: {
+            oneOf: [
+                { type: 'string' },
+                { type: 'array', items: { type: 'string' } },
+                { type: 'null' }
+            ],
+            description: 'The matching IMDb ID(s) for the entry. Can be a single ID or an array of IDs',
+            example: 'tt9288776'
+        },
+        mal_id: {
+            oneOf: [
+                { type: 'integer' },
+                { type: 'array', items: { type: 'integer' } },
+                { type: 'null' }
+            ],
+            description: 'The matching MyAnimeList ID(s) for the entry. Can be a single ID or an array of IDs',
+            example: [34915, 36186, 36529]
+        },
+        tmdb_movie_id: {
+            oneOf: [
+                { type: 'integer' },
+                { type: 'array', items: { type: 'integer' } },
+                { type: 'null' }
+            ],
+            description: 'The matching TMDB ID(s) for the movie entry. Can be a single ID or an array of IDs',
+            example: 532067
+        },
+        tmdb_show_id: {
+            oneOf: [
+                { type: 'integer' },
+                { type: 'array', items: { type: 'integer' } },
+                { type: 'null' }
+            ],
+            description: 'The matching TMDB ID(s) for the show entry. Can be a single ID or an array of IDs',
+            example: 83095
+        },
+        tvdb_id: {
+            type: 'integer',
+            nullable: true,
+            description: 'The matching TVDB ID for the show entry',
+            example: 333234
+        },
+        tvdb_mappings: {
+            type: 'object',
+            nullable: true,
+            additionalProperties: { type: 'string' },
+            description: 'A dictionary mapping TVDB seasons to episode patterns',
+            example: {
+                's0': 'e1',
+                's1': 'e1-e13'
+            }
+        }
+    },
+    example: {
+        anidb_id: 12820,
+        anilist_id: 98310,
+        imdb_id: 'tt9288776',
+        mal_id: [34915, 36186, 36529],
+        tmdb_movie_id: null,
+        tmdb_show_id: 83095,
+        tvdb_id: 333234,
+        tvdb_mappings: {
+            's0': 'e1',
+            's1': 'e1-e13'
+        }
+    }
+});
+
+
+
+export const ErrorSchemas = {
+    InternalServerError: z.object({
+        error: z.string(),
+    }).openapi({
+        description: 'Internal server error',
+    }),
+
+    NotFoundError: z.object({
+        error: z.string(),
+    }).openapi({
+        description: 'Resource not found',
+    }),
+
+    BadRequestError: z.object({
+        error: z.string(),
+    }).openapi({
+        description: 'Bad request error',
+    })
+};
